Some final cleanup
